Create a FAQ to describe difference among notary, TUF, Notary Project specification, and notation (#328)
This PR adds a new section in faq.md to clarify the difference between Notary, Notary v2, Notation, TUF, and Notary Project. It fixes #323 

I haven't included a link to the `glossary page` in the last line because PR #326 is still under review and has not been merged yet. I will update this PR with the link once #326 is merged.

---------

Signed-off-by: Roseline <roselynbassey23@gmail.com>
Co-authored-by: Yi Zha <yizha1@microsoft.com>