chroe: add security audit report 2025 to the website and refactor the security audit page (#439)
Add security audit report to the website. It was only available in the
specification repo.

Preview link:
https://deploy-preview-439--notarydev.netlify.app/security-audit/.

I also refactor the security audit page since it was not able to show
multiple security audit links. Now only one small issue is that the 2025
audit report is not aligned with others. This will take a longer time to
improve the front-end code. It does not impact readers' experience so we
can merge it as the first iteration.

---------

Signed-off-by: Feynman Zhou <feynmanzhou@microsoft.com>